Move test support into snitch tools
The test support cases are extremely handy in not just tests but also
seeding.
This change makes these modules available at runtime (dev env).
