--- apps/snitch_core/dots/seed.exs
@@ -1,3 +1,0 @@
-alias Snitch.Seed.{Orders, Users}
-alias Snitch.Data.Schema.{Variant, Order, User}
-alias Snitch.Repo
