+++ apps/snitch_core/lib/core/tools/helpers/shipment.ex
@@ -1,16 +1,17 @@
-defmodule Snitch.ShipmentCase do
+defmodule Snitch.Tools.Helper.Shipment do
   @moduledoc """
   Test helpers to insert shipping related schemas.
   """
 
+  alias Ecto.DateTime
   alias Snitch.Data.Model.ShippingMethod
   alias Snitch.Data.Schema.ShippingCategory
   alias Snitch.Repo
 
   @shipping_category %{
     name: nil,
-    inserted_at: Ecto.DateTime.utc(),
-    updated_at: Ecto.DateTime.utc()
+    inserted_at: DateTime.utc(),
+    updated_at: DateTime.utc()
   }
 
   @doc """
