+++ apps/snitch_core/test/data/model/zone/zone_test.exs
@@ -3,7 +3,7 @@ defmodule Snitch.Data.Model.ZoneTest do
   use Snitch.DataCase
 
   import Snitch.Factory
-  import Snitch.ZoneCase
+  import Snitch.Tools.Helper.Zone
 
   alias Snitch.Data.Model.{CountryZone, StateZone}
   alias Snitch.Domain.Zone
